requirements_afm.txt
bottle
webtest
SQLAlchemy
jsonschema
nose
coverage
dateutils
psycopg2
